# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AskWorkspaceQuestionRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskWorkspaceQuestionRequest",
        "type": "object"
      },
...
      "Citation": {
        "additionalProperties": false,
        "properties": {
          "citationId": {
            "format": "uuid",
            "title": "Citationid",
            "type": "string"
          },
          "documentVersionId": {
            "format": "uuid",
            "title": "Documentversionid",
            "type": "string"
          },
          "locator": {
            "discriminator": {
              "mapping": {
                "pdf_page": "#/components/schemas/PdfPageLocator",
                "text_lines": "#/components/schemas/TextLineLocator"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextLineLocator"
              },
              {
                "$ref": "#/components/schemas/PdfPageLocator"
              }
            ],
            "title": "Locator"
          },
          "quote": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          },
          "sourceBlockId": {
            "format": "uuid",
            "title": "Sourceblockid",
            "type": "string"
          }
        },
        "required": [
          "citationId",
          "documentVersionId",
          "locator",
          "quote",
          "sourceBlockId"
        ],
        "title": "Citation",
        "type": "object"
      },
      "CitationContext": {
        "additionalProperties": false,
        "properties": {
          "citationId": {
            "format": "uuid",
            "title": "Citationid",
            "type": "string"
          },
          "fileName": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "locatorLabel": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Locatorlabel",
            "type": "string"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "passageAfter": {
            "maxLength": 221,
            "title": "Passageafter",
            "type": "string"
          },
          "passageBefore": {
            "maxLength": 221,
            "title": "Passagebefore",
            "type": "string"
          }
        },
        "required": [
          "citationId",
          "fileName",
          "locatorLabel",
          "observedAt",
          "passageAfter",
          "passageBefore"
        ],
        "title": "CitationContext",
        "type": "object"
      },
      "ContextManifest": {
        "additionalProperties": false,
        "properties": {
          "activeClaimIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 12,
            "title": "Activeclaimids",
            "type": "array"
          },
          "activeTurnIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 4,
            "title": "Activeturnids",
            "type": "array"
          },
          "candidateChunkIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 40,
            "title": "Candidatechunkids",
            "type": "array"
          },
          "contextPolicyVersion": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Contextpolicyversion",
            "type": "string"
          },
          "excludedChunks": {
            "items": {
              "$ref": "#/components/schemas/ExcludedContextChunk"
            },
            "maxItems": 40,
            "title": "Excludedchunks",
            "type": "array"
          },
          "includedChunkIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 12,
            "title": "Includedchunkids",
            "type": "array"
          },
          "originalMessageHash": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Originalmessagehash",
            "type": "string"
          },
          "selectionOrdering": {
            "const": "retrieval_rank_then_source_cap",
            "title": "Selectionordering",
            "type": "string"
          },
          "snapshotId": {
            "format": "uuid",
            "title": "Snapshotid",
            "type": "string"
          },
          "sourceCounts": {
            "items": {
              "$ref": "#/components/schemas/ContextSourceCount"
            },
            "maxItems": 40,
            "title": "Sourcecounts",
            "type": "array"
          },
          "tokenBudget": {
            "exclusiveMinimum": 0.0,
            "maximum": 16000.0,
            "title": "Tokenbudget",
            "type": "integer"
          }
        },
        "required": [
          "activeClaimIds",
          "activeTurnIds",
          "candidateChunkIds",
          "contextPolicyVersion",
          "excludedChunks",
          "includedChunkIds",
          "originalMessageHash",
          "selectionOrdering",
          "snapshotId",
          "sourceCounts",
          "tokenBudget"
        ],
        "title": "ContextManifest",
        "type": "object"
      },
      "ContextSourceCount": {
        "additionalProperties": false,
        "properties": {
          "candidateChunkCount": {
            "minimum": 0.0,
            "title": "Candidatechunkcount",
            "type": "integer"
          },
          "documentVersionId": {
            "format": "uuid",
            "title": "Documentversionid",
            "type": "string"
          },
          "includedChunkCount": {
            "minimum": 0.0,
            "title": "Includedchunkcount",
            "type": "integer"
          }
        },
        "required": [
          "candidateChunkCount",
          "documentVersionId",
          "includedChunkCount"
        ],
        "title": "ContextSourceCount",
        "type": "object"
      },
...
      "ExcludedContextChunk": {
        "additionalProperties": false,
        "properties": {
          "chunkId": {
            "format": "uuid",
            "title": "Chunkid",
            "type": "string"
          },
          "reasonCode": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Reasoncode",
            "type": "string"
          }
        },
        "required": [
          "chunkId",
          "reasonCode"
        ],
        "title": "ExcludedContextChunk",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublishedAnswerView": {
        "additionalProperties": false,
        "properties": {
          "answerId": {
            "format": "uuid",
            "title": "Answerid",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Citations",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/PublishedClaim"
            },
            "maxItems": 3,
            "minItems": 1,
            "title": "Claims",
            "type": "array"
          },
          "funnel": {
            "$ref": "#/components/schemas/EvidenceFunnel"
          },
          "revision": {
            "$ref": "#/components/schemas/ExtentRevision"
          },
          "terminal": {
            "$ref": "#/components/schemas/PublishedTerminal"
          },
          "traceId": {
            "format": "uuid",
            "title": "Traceid",
            "type": "string"
          }
        },
        "required": [
          "answerId",
          "citations",
          "claims",
          "funnel",
          "revision",
          "terminal",
          "traceId"
        ],
        "title": "PublishedAnswerView",
        "type": "object"
      },
...
      "QueryExecution": {
        "additionalProperties": false,
        "properties": {
          "acknowledgementCopyKey": {
            "const": "evidence_acknowledgement",
            "title": "Acknowledgementcopykey",
            "type": "string"
          },
          "contextManifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextManifest"
              },
              {
                "type": "null"
              }
            ]
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "stages": {
            "items": {
              "enum": [
                "accepted",
                "authorizing",
                "interpreting",
                "cache_check",
                "contextualizing",
                "retrieving",
                "composing",
                "structural_validation",
                "verifying",
                "publishing",
                "complete"
              ],
              "type": "string"
            },
            "maxItems": 11,
            "minItems": 3,
            "title": "Stages",
            "type": "array"
          },
          "view": {
            "$ref": "#/components/schemas/PublishedAnswerView"
          }
        },
        "required": [
          "acknowledgementCopyKey",
          "contextManifest",
          "replayed",
          "stages",
          "view"
        ],
        "title": "QueryExecution",
        "type": "object"
      },
      "SampleWorkspaceProjection": {
        "additionalProperties": false,
        "properties": {
          "citationContexts": {
            "items": {
              "$ref": "#/components/schemas/CitationContext"
            },
            "maxItems": 6,
            "title": "Citationcontexts",
            "type": "array"
          },
          "execution": {
            "$ref": "#/components/schemas/QueryExecution"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceSummary"
          }
        },
        "required": [
          "citationContexts",
          "execution",
          "question",
          "workspace"
        ],
        "title": "SampleWorkspaceProjection",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceApprovedClaimView": {
        "additionalProperties": false,
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceEvidencePassageView"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Citations",
            "type": "array"
          },
          "claimId": {
            "format": "uuid",
            "title": "Claimid",
            "type": "string"
          },
          "relation": {
            "enum": [
              "fact",
              "change",
              "conflict",
              "unclear"
            ],
            "title": "Relation",
            "type": "string"
          },
          "text": {
            "maxLength": 800,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "citations",
          "claimId",
          "relation",
          "text",
          "value"
        ],
        "title": "WorkspaceApprovedClaimView",
        "type": "object"
      },
      "WorkspaceErrorView": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "authentication_required",
              "idempotency_conflict",
              "invalid_folder_url",
              "origin_rejected",
              "rate_limit_unavailable",
              "rate_limited",
              "retrieval_unavailable",
              "workspace_not_retryable",
              "workspace_not_ready",
              "workspace_not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 280,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "reasonCode": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoncode"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "WorkspaceErrorView",
        "type": "object"
      },
      "WorkspaceEvidencePassageView": {
        "additionalProperties": false,
        "properties": {
          "blockId": {
            "format": "uuid",
            "title": "Blockid",
            "type": "string"
          },
          "driveFileId": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Drivefileid",
            "type": "string"
          },
          "endExclusiveInBlock": {
            "exclusiveMinimum": 0.0,
            "title": "Endexclusiveinblock",
            "type": "integer"
          },
          "exactQuote": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Exactquote",
            "type": "string"
          },
          "lineStartOneBased": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linestartonebased"
          },
          "normalizedValue": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalizedvalue"
          },
          "originKind": {
            "enum": [
              "pdf_page",
              "text_lines"
            ],
            "title": "Originkind",
            "type": "string"
          },
          "pageIndexZeroBased": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pageindexzerobased"
          },
          "path": {
            "items": {
              "maxLength": 1024,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 2,
            "title": "Path",
            "type": "array"
          },
          "printedPageLabel": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printedpagelabel"
          },
          "rawValue": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rawvalue"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "support",
                  "before",
                  "after",
                  "left",
                  "right"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "sourceName": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Sourcename",
            "type": "string"
          },
          "startInBlock": {
            "minimum": 0.0,
            "title": "Startinblock",
            "type": "integer"
          }
        },
        "required": [
          "blockId",
          "driveFileId",
          "endExclusiveInBlock",
          "exactQuote",
          "lineStartOneBased",
          "normalizedValue",
          "originKind",
          "pageIndexZeroBased",
          "path",
          "printedPageLabel",
          "rawValue",
          "role",
          "sourceName",
          "startInBlock"
        ],
        "title": "WorkspaceEvidencePassageView",
        "type": "object"
      },
...
      "WorkspaceQuestionResultView": {
        "additionalProperties": false,
        "properties": {
          "answerId": {
            "format": "uuid",
            "title": "Answerid",
            "type": "string"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceApprovedClaimView"
            },
            "maxItems": 200,
            "title": "Claims",
            "type": "array"
          },
          "coverageGapReasons": {
            "items": {
              "enum": [
                "processing",
                "failed",
                "unsupported",
                "inaccessible",
                "capped",
                "unknown_branch",
                "unstable",
                "unsafe_to_parse"
              ],
              "type": "string"
            },
            "maxItems": 8,
            "title": "Coveragegapreasons",
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "generationStatus": {
            "enum": [
              "not_configured",
              "failed",
              "completed"
            ],
            "title": "Generationstatus",
            "type": "string"
          },
          "message": {
            "maxLength": 280,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "passages": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceEvidencePassageView"
            },
            "maxItems": 6,
            "title": "Passages",
            "type": "array"
          },
          "policyVersion": {
            "enum": [
              "retrieval-policy-v1",
              "publication-policy-v1",
              "clarification-policy-v1",
              "exhaustive-extraction-policy-v1",
              "exhaustive-premium-policy-v1",
              "source-state-policy-v1",
              "structured-analysis-policy-v1"
            ],
            "title": "Policyversion",
            "type": "string"
          },
          "question": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "questionId": {
            "format": "uuid",
            "title": "Questionid",
            "type": "string"
          },
          "status": {
            "enum": [
              "evidence_retrieved",
              "evidence_supported",
              "changed",
              "conflict",
              "insufficient",
              "coverage_limited"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "answerId",
          "claims",
          "coverageGapReasons",
          "createdAt",
          "generationStatus",
          "message",
          "passages",
          "policyVersion",
          "question",
          "questionId",
          "status"
        ],
        "title": "WorkspaceQuestionResultView",
        "type": "object"
      },
      "WorkspaceSource": {
        "additionalProperties": false,
        "properties": {
          "documentVersionId": {
            "format": "uuid",
            "title": "Documentversionid",
            "type": "string"
          },
          "evaluated": {
            "title": "Evaluated",
            "type": "boolean"
          },
          "fileName": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "observedAt": {
            "format": "date-time",
            "title": "Observedat",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "selected": {
            "title": "Selected",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "ready",
              "unsupported"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "documentVersionId",
          "evaluated",
          "fileName",
          "observedAt",
          "selected",
          "status"
        ],
        "title": "WorkspaceSource",
        "type": "object"
      },
...
      "WorkspaceSummary": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revisionLabel": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revisionlabel"
          },
          "sampleLabel": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Samplelabel",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceSource"
            },
            "maxItems": 100,
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "name",
          "revisionLabel",
          "sampleLabel",
          "sources"
        ],
        "title": "WorkspaceSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Extent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/demo/questions": {
      "post": {
        "operationId": "ask_demo_question",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskWorkspaceQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceQuestionResultView"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceErrorView"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceErrorView"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceErrorView"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Ask an anonymous question about the prepared sample",
        "tags": [
          "demo"
        ]
      }
    },
    "/api/v1/demo/workspace": {
      "get": {
        "operationId": "get_demo_workspace",
        "parameters": [
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleWorkspaceProjection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the interactive Alder Peak sample workspace",
        "tags": [
          "demo"
        ]
      }
    },
...
  }
}